+++ src/Price.php
@@ -129,7 +129,7 @@ class Price implements XmlSerializable, XmlDeserializable
         $writer->write([
             [
                 'name'       => Schema::CBC . 'PriceAmount',
-                'value'      => NumberFormatter::format($this->priceAmount),
+                'value'      => NumberFormatter::format($this->priceAmount, 2),
                 'attributes' => [
                     'currencyID' => Generator::$currencyID
                 ]